# docs/swagger.yaml
basePath: /api/
definitions:
...
  domain.CreateCustomerRequest:
    properties:
      alamat:
        maxLength: 1000
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        minLength: 12
        type: string
    required:
    - alamat
    - email
    - name
    - phone
    type: object
...
  domain.CreateProductRequest:
    properties:
      category_id:
        type: string
      image:
        type: string
      name:
        type: string
      qty:
        type: string
    required:
    - category_id
    - image
    - name
    - qty
    type: object
...
  domain.ErrorMessage:
    properties:
      error:
        type: boolean
      message:
        type: string
      statusCode:
        type: integer
    type: object
...
  domain.Response:
    properties:
      data: {}
      message:
        type: string
      statusCode:
        type: integer
    type: object
...
  domain.UpdateCustomerRequest:
    properties:
      alamat:
        maxLength: 1000
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        minLength: 12
        type: string
    required:
    - alamat
    - email
    - id
    - name
    - phone
    type: object
...
  domain.UpdateProductRequest:
    properties:
      category_id:
        type: string
      id:
        type: string
      image:
        type: string
      name:
        type: string
      qty:
        type: string
    required:
    - category_id
    - image
    - name
    - qty
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: REST API for FiberInventory
  title: FiberInventory
  version: "1.0"
paths:
  /api/customer:
    get:
      description: Retrieve the results for each customer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Get customer results
      tags:
      - Customer
...
  /customer/{id}:
    delete:
      description: Delete a specific customer
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Delete customer
      tags:
      - Customer
    get:
      description: Retrieve the result for a specific customer
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Get customer result
      tags:
      - Customer
    put:
      description: Update a specific customer
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      - description: Customer Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateCustomerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Update Customer
      tags:
      - Customer
  /customer/create:
    post:
      consumes:
      - application/json
      description: Create Customer
      parameters:
      - description: Customer information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateCustomerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: createCustomer to the application
      tags:
      - Customer
  /customer/hello:
    get:
      description: Return a greeting message to the customer
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Greet the Customer
      tags:
      - Customer
  /product:
    get:
      description: Retrieve the results for each product
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Get product results
      tags:
      - Product
  /product/{id}:
    delete:
      description: Delete a specific product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Delete product
      tags:
      - Product
    get:
      description: Retrieve the result for a specific product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Get product result
      tags:
      - Product
    put:
      description: Update a specific product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - Product
  /product/create:
    post:
      consumes:
      - application/json
      description: Create Product
      parameters:
      - description: Product information
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/domain.CreateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: createProduct to the application
      tags:
      - product
  /product/hello:
    get:
      description: Return a greeting message to the product
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Greet the Product
      tags:
      - Product
...
securityDefinitions:
  BearerAuth:
    in: Header
    name: Authorization
    type: apiKey
swagger: "2.0"